Setup for calling CustomAPI.
This is a Deviare feature that will call a named function in the NativePlugin DLL.

The tricky part was getting the path correct, so that the dll could be found.  With that in place, it's straightforward to call the native routines directly.

We can use this to turn the wireframe on and off.