Add Registration Free com manifest file.
Having to register the DeviareCOM.dll is a pain. This adds an InvisibleWalls.manifest file that specifies the DeviareCOM to be used.

This works for x86, both Debug/Release.

For reasons that are not clear, this needs to be embedded in the Application settings, and does not work if it is standalone beside the exe.  (Most likely the app .config overrides the standalone.)