Fix Nektra.Deviare2.dll hint path.
I have no idea why this did not get updated and required directly modifying the csproj xml file.

This adds the reference to the Deviare2.dll needed for the C# to compile.

It also works if the "Deviare2 Dependencies" is added to the Reference Paths in the Properties, but that changes no files, so was not saved in git.